+++ gui/wxpython/main_window/frame.py
@@ -2010,12 +2010,19 @@ class GMFrame(wx.Frame):
         win.Show()
 
     def OnAnimationTool(self, event=None, cmd=None):
-        """Launch Animation tool. See OnIClass documentation."""
-        from animation.frame import AnimationFrame
+        """Launch Animation tool. See OnIClass documentation.
 
-        frame = AnimationFrame(parent=self, giface=self._giface)
-        frame.CentreOnScreen()
-        frame.Show()
+        :return: the animation panel, so that a caller can load data into it
+        """
+        from animation.panels import AnimationPanel
+
+        animation_panel = AnimationPanel(
+            parent=self, giface=self._giface, statusbar=self.statusbar, dockable=True
+        )
+        animation_panel.SetUpPage(self, self.mainnotebook)
+
+        # add animation panel to notebook and make it current
+        self.mainnotebook.AddPage(animation_panel, _("Animation Tool"))
 
         tree = self.GetLayerTree()
         if tree:
@@ -2034,7 +2041,9 @@ class GMFrame(wx.Frame):
                 layer.name = ",".join(rasters)
                 layer.cmd = ["d.rast", "map="]
                 layerList.AddLayer(layer)
-                frame.SetAnimations([layerList, None, None, None])
+                animation_panel.SetAnimations([layerList, None, None, None])
+
+        return animation_panel
 
     def OnTimelineTool(self, event=None, cmd=None):
         """Launch Timeline Tool"""
